Fix BOOLEAN const pushdown to not include quotes
When a filter with a `BOOLEAN` constant is pushed down to attached MySQL
DB, current implementation treats it as a `VARCHAR` and incidentally
transforms the filter:

```sql
a = TRUE
```

into:

```sql
a = 'TRUE'
```

that, when executed, is always false for `BOOLEAN` values.

This change makes the handling for `BOOLEAN` constant values in
pushdowns to be the same same as numeric values, so they are passed
without additional quotes.

Additionally it fixes the typo (missed space) when generating an
`UPDATE` query.

Testing: pushdown test is updated to output the actual column value
instead of the `COUNT` so it can catch this problem; scan test is
updated to cover the boolean roundrip on insert/update.