Allow setting MySQL session time zone
In MySQL `TIMESTAMP` columns have the following logic:

> MySQL converts TIMESTAMP values from the current time zone to UTC for
> storage, and back from UTC to the current time zone for retrieval.

This change adds a new setting `mysql_session_time_zone` (`VARCHAR`,
empty by default) that is used to set the `TIME_ZONE` option in all
newly opened connections to MySQL server.

There is no attempt to synchronize this new setting with the DuckDB's
`TimeZone` setting for the following reasons:

1. time zones handling requires ICU extension that, in theory, may no be
   available to the scanner extension

2. user may want to not set this setting and rely on the MySQL server
   time zone setting instead (default behaviour), thus using DuckDB's
   time zone value for MySQL session time zone may be undesired

Testing: new test is added that checks that new setting is effective in
newly opened connections.