Fix DATE const handling in UPDATE queries
When `UPDATE` queries are generated for a few types (including `DATE`)
the incoming constant (in `SET` clause) is converted to `VARCHAR` using
`Value#ToSQLString()` function. This call adds Postgres'-like cast
suffix to the literal (like `::DATE`) and this syntax is not supported
by MySQL/MariaDB.

This PR changes the constant passing in `UPDATE` using the same logic
that is used for filter pushdowns.

Testing: existing test is updated to cover `UPDATE`s with `DATE`
literals.

Fixes: #143