Fix bind buffer usage with long VARCHARs
This change fixes the problem with the misuse of the
`mysql_stmt_fetch_column` API (introduced in `1.4.1`) when this call was
used to fetch the remaining part of the long `VARCHAR` string. It
appeared that MySQL connector does not allow to resize the buffer
without advancing to the next row. The problem reported in #170 was
reproducible when 2 or more long `VARCHARs` are present on the same row.

This PR adds a separate buffer for long `VARCHARs`, otherwise keeping
the fetching logic intact.

Testing: existing test updated to cover fetching the rows with multiple
long strings.

Fixes: #170