Use libmariadb client library instead of libmysql
This PR swithes the dependency from `libmysql` to `libmariadb`.

There are no functional changes. SSL options with `libmariadb` are
slightly different. Extension SSL options are kept the same and are
translated into corresponding `libmariadb` options.

Testing: test suite was run with both MySQL and MariaDB servers. SSL
options are checked with MySQL with server SSL on and off.