Do not convert zero TIME to NULL (1.4)
This is a backport of the PR #177 to `v1.4-andium` stable branch.

In MySQL/MariaDB C connector the `DATE`, `TIME`, `DATETIME` and
`TIMESTAMP` values are fetched using the same `MYSQL_TIME` struct:

```c
typedef struct st_mysql_time
{
  unsigned int  year, month, day, hour, minute, second;
  unsigned long second_part;
  my_bool       neg;
  enum enum_mysql_timestamp_type time_type;
} MYSQL_TIME;
```

The scanner has implemented the following note from the connector
documentation:

> "Zero" date or time values used through Connector/ODBC are converted automatically to NULL because ODBC cannot handle such values.

It appeared that outside of ODBC (that is not used by the scanner) MySQL
treats "zero" values as `NULL`s for `DATE`, `DATETIME` and `TIMESTAMP`
values, but not for `TIME` values.

This change removes the logic that treats `00:00:00` `TIME` values as
`NULL`s. So '00:00:00' can now be returned to client.

Note, that MySQL `TIME` can be mapped either to DuckDB `TIME` or
`VARCHAR` types (depending on `mysql_time_as_time` option added
in #142), this change applies to both types.

Testing: new tests added for "zero" temporal values.

Ref: duckdblabs/duckdb-internal#6614