Do not fetch full result set on close
Unlike the Simple API, the Prepared Statement API does not require the
client (scanner) to read all the results before closing the result set.
Instead it is doing that under the hood in `mysql_stmt_close` call.
With large result set this can be time consuming and for end-user looks
like the query is hanging.

This PR re-implements the "reconnect+kill" approach from #136 for
prepared statements. When the query is run in streaming mode (see
notes on streaming in [this comment](https://github.com/duckdb/duckdb-mysql/issues/55#issuecomment-3607311166))
there are 2 scenarios:

1. the subsequent query is run on the same connection without reading
   all the result from the previous query

2. the running query is interrupted (using `Connection->Interrrup()`
   or its client-specific wrapper) from another thread

In both cases the second connection to MySQL server is established and
`KILL QUERY <ID>` command is run on that connection.

Unlike the #136, this change does not swap the underlying MySQL
connection to a new one:

 - the cleanup in `mysql_stmt_close` after `KILL QUERY` leaves the MySQL
   connection in a valid state
 - when the query is interrupted (that causes the transaction abort) or
   when the auto-commit is enabled - the following queries on the same
   DuckDB connection will open new MySQL connections
 - so the actual connection reuse after the query killed on this
   connection happens only when the result set is abandoned (not
   exhausted) and the subsequent query is run on the same connection in
   the same transaction.

Note, that language clients may have their own restrictions on result
set streaming, for example, in JDBC the streaming is disabled by
default.

Testing: the query abandon (or interrupt) logic is not supported by
`sqllogictest` test suite. The scenarios above are tested manually with
TPC-H data.